# fastapi/测试脚本/v2测试脚本/test/导出接口文档脚本/fastapi_complete_api_spec.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "用户系统API",
    "description": "模块化的用户管理系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v2/document_manager/folders": {
      "post": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "创建文件夹",
        "description": "创建新文件夹\n\n- **name**: 文件夹名称（1-100字符，不能包含特殊字符）\n- **parent_id**: 父文件夹ID（可选，不填表示根目录）\n\n限制：\n- 同一位置不能有同名文件夹\n- 最多支持3层文件夹结构",
        "operationId": "create_folder_api_v2_document_manager_folders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/document_manager/folders/tree": {
      "get": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "获取文件夹树",
        "description": "获取当前用户的文件夹树形结构\n\n返回完整的文件夹层级关系，包含每个文件夹下的文档数量",
        "operationId": "get_folder_tree_api_v2_document_manager_folders_tree_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FolderTreeResponse"
                  },
                  "type": "array",
                  "title": "Response Get Folder Tree Api V2 Document Manager Folders Tree Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/document_manager/folders/{folder_id}": {
      "delete": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "删除文件夹",
        "description": "删除文件夹\n\n注意：\n- 只能删除空文件夹（无子文件夹和文档）\n- 删除操作不可恢复",
        "operationId": "delete_folder_api_v2_document_manager_folders__folder_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Folder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/document_manager/documents": {
      "post": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "创建文档",
        "description": "创建新文档\n\n- **title**: 文档标题（1-200字符）\n- **content**: 文档内容（Markdown格式，可选）\n- **summary**: 简短摘要（最多500字符，可选）\n- **folder_id**: 所属文件夹ID（可选，不填表示根目录）\n- **file_type**: 文件类型（md或pdf，默认md）\n\n限制：\n- 同一文件夹下不能有同名文档\n- 新创建的文档默认为草稿状态",
        "operationId": "create_document_api_v2_document_manager_documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "获取文档列表",
        "description": "获取文档列表（分页）\n\n参数：\n- **folder_id**: 文件夹ID筛选（可选）\n  - 不填：获取所有文档\n  - 0：获取根目录下的文档\n  - 其他数字：获取指定文件夹下的文档\n- **page**: 页码（从1开始）\n- **page_size**: 每页数量（1-100）\n\n返回：\n- 文档列表（按更新时间倒序）\n- 分页信息",
        "operationId": "get_documents_list_api_v2_document_manager_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "文件夹ID，不填获取所有文档，0表示根目录",
              "title": "Folder Id"
            },
            "description": "文件夹ID，不填获取所有文档，0表示根目录"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListWithPaginationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/document_manager/documents/{doc_id}": {
      "get": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "获取文档详情",
        "description": "获取指定文档的详细信息\n\n包含完整的文档内容、状态、所属文件夹等信息",
        "operationId": "get_document_api_v2_document_manager_documents__doc_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "更新文档",
        "description": "更新文档信息\n\n可以更新：\n- 标题\n- 内容\n- 摘要\n- 所属文件夹\n\n注意：\n- 只能更新自己的文档\n- 更新后文档状态可能需要重新审核",
        "operationId": "update_document_api_v2_document_manager_documents__doc_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "删除文档",
        "description": "删除文档\n\n注意：\n- 删除操作不可恢复\n- 会同时删除相关的物理文件\n- 如果文档已发布，会从技术广场移除",
        "operationId": "delete_document_api_v2_document_manager_documents__doc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/document_manager/stats": {
      "get": {
        "tags": [
          "v2-document_manager"
        ],
        "summary": "获取统计信息",
        "description": "获取用户的文档统计信息\n\n包含：\n- 总文档数\n- 各状态文档数量\n- 文件夹数量",
        "operationId": "get_user_stats_api_v2_document_manager_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v2/interaction/documents/{document_id}/stats": {
      "get": {
        "tags": [
          "v2-interaction"
        ],
        "summary": "Get Document Stats",
        "description": "获取文档互动统计",
        "operationId": "get_document_stats_api_v2_interaction_documents__document_id__stats_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "文档标题"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "文档内容（MD格式）"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "简短摘要"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id",
            "description": "所属文件夹ID"
          },
          "file_type": {
            "$ref": "#/components/schemas/FileTypeEnum",
            "description": "文件类型",
            "default": "md"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DocumentCreateRequest",
        "description": "创建文档请求"
      },
...
      "DocumentListWithPaginationResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/app__modules__v2__document_manager__schemas__DocumentListResponse"
            },
            "type": "array",
            "title": "Documents"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "documents",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "DocumentListWithPaginationResponse",
        "description": "带分页的文档列表响应"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "file_type": {
            "$ref": "#/components/schemas/FileTypeEnum"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatusEnum"
          },
          "publish_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Time"
          },
          "review_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Message"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "file_path",
          "file_type",
          "file_size",
          "summary",
          "status",
          "publish_time",
          "review_message",
          "folder_id",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse",
        "description": "文档响应"
      },
      "DocumentStatusEnum": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "review_failed"
        ],
        "title": "DocumentStatusEnum",
        "description": "文档状态枚举"
      },
      "DocumentUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "文档标题"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "文档内容"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "简短摘要"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id",
            "description": "所属文件夹ID"
          }
        },
        "type": "object",
        "title": "DocumentUpdateRequest",
        "description": "更新文档请求"
      },
...
      "FileTypeEnum": {
        "type": "string",
        "enum": [
          "md",
          "pdf"
        ],
        "title": "FileTypeEnum",
        "description": "文件类型枚举"
      },
...
      "FolderCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "文件夹名称"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "父文件夹ID，不填表示根目录"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FolderCreateRequest",
        "description": "创建文件夹请求"
      },
      "FolderResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "parent_id",
          "level",
          "created_at",
          "updated_at"
        ],
        "title": "FolderResponse",
        "description": "文件夹响应"
      },
      "FolderTreeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/FolderTreeResponse"
            },
            "type": "array",
            "title": "Children",
            "default": []
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "level"
        ],
        "title": "FolderTreeResponse",
        "description": "文件夹树形结构响应"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InteractionStats": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "favorite_count": {
            "type": "integer",
            "title": "Favorite Count"
          },
          "comment_count": {
            "type": "integer",
            "title": "Comment Count"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "like_count",
          "favorite_count",
          "comment_count",
          "updated_at"
        ],
        "title": "InteractionStats",
        "description": "互动统计模型"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SuccessResponse",
        "description": "成功响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__modules__v2__document_manager__schemas__DocumentListResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "file_type": {
            "$ref": "#/components/schemas/FileTypeEnum"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatusEnum"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "file_type",
          "file_size",
          "status",
          "folder_id",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentListResponse",
        "description": "文档列表响应"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}